LayoutTests/accessibility/mac/element-level-expected.txt

blockquote level 1
blockquote level 2
heading level 1

heading level 2

table heading level 1
table heading level 2
table body level 2
This tests elements' accessibility level value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS body.childAtIndex(0).childAtIndex(0).numberAttributeValue('AXBlockQuoteLevel') is 1
PASS body.childAtIndex(1).childAtIndex(0).numberAttributeValue('AXBlockQuoteLevel') is 2
PASS body.childAtIndex(2).numberAttributeValue('AXBlockQuoteLevel') is 0
PASS body.childAtIndex(2).numberAttributeValue('AXValue') is 1
PASS body.childAtIndex(3).numberAttributeValue('AXValue') is 2
PASS body.childAtIndex(4).numberAttributeValue('AXTableLevel') is 1
PASS body.childAtIndex(4).childAtIndex(1).childAtIndex(0).childAtIndex(0).numberAttributeValue('AXTableLevel') is 2
PASS body.childAtIndex(0).numberAttributeValue('AXTableLevel') is 0
PASS successfullyParsed is true

TEST COMPLETE
